Bust menucommand_handlers into module (directory) and individual files Fixes #25
