+++ calvincTools/cMenu/constants.py
@@ -1,0 +1,7 @@
+# DONE: put in class?   In module-level for now, but could be in cMenu class or separate class if needed
+# cMenu-related constants
+_SCRN_menuBTNWIDTH:int = 250
+_SCRN_menuDIVWIDTH:int = 40
+_NUM_menuBUTTONS:int = 20
+_NUM_menuBUTNCOLS:int = 2
+_NUM_menuBTNperCOL: int = int(_NUM_menuBUTTONS/_NUM_menuBUTNCOLS)
