+++ calvincTools/cMenu/menucommand_handlers/__init__.py
@@ -1,0 +1,8 @@
+
+Nochoice = {'---': None}    # only needed for combo boxes, not datalists
+
+from .changePassword import changePassword
+from .FormBrowse import FormBrowse
+from .loadExternalWebPage import loadExternalWebPage
+from .OpenTable import OpenTable
+from .RunSQL import cMRunSQL
